requirements.txt
openai
tiktoken
python-dotenv
rich
discord.py
pandas
scipy
pyaudio
speechrecognition
pyttsx3
keyboard
playsound
pytz
customtkinter
pillow